+++ contract-macro/tests/compile-pass/src/data_driver/mod.rs
@@ -1,0 +1,5 @@
+pub mod module_local_const;
+pub mod module_local_impl_only_use;
+pub mod module_local_private_type_alias;
+pub mod module_local_struct_field;
+pub mod module_local_type_alias;